+++ src/styles/Book.module.css
@@ -1,0 +1,189 @@
+.bookcardcontainer {
+  background-color: #fff;
+  display: grid;
+  grid-template-columns: 2fr 1fr 1.5fr;
+  justify-content: space-between;
+  border-radius: 5px;
+  padding: 2rem;
+}
+
+.bookinfocontainer {
+  display: grid;
+  grid-template-columns: 1fr 1fr;
+  gap: 2rem;
+}
+
+.column {
+  display: flex;
+  flex-direction: column;
+  gap: 0.7rem;
+}
+
+.maininfocontainer {
+  display: flex;
+  flex-direction: column;
+  gap: 0.7rem;
+}
+
+.categorycontainer,
+.completedWord,
+.currentChapterTitle {
+  color: #121212;
+  opacity: 0.5;
+}
+
+.titlecontainer {
+  font-size: 1.5rem;
+}
+
+.authorcontainer {
+  color: #0290ff;
+}
+
+.buttonscontainer {
+  display: flex;
+  gap: 1rem;
+}
+
+.button,
+.button1 {
+  border: none;
+  border-left: solid 1px rgba(128, 128, 128, 0.418);
+  background-color: transparent;
+  color: #0290ff;
+  padding: 0 1rem;
+  cursor: pointer;
+}
+
+.button1 {
+  border: none;
+  padding-left: 0;
+}
+
+.completed {
+  display: flex;
+  gap: 1rem;
+  margin-right: -1rem;
+  margin-left: 30rem;
+  margin-top: -10rem;
+  padding-left: 1rem;
+}
+
+.progressCompleted {
+  display: flex;
+  flex-direction: column;
+  gap: 0.7rem;
+}
+
+.percentage {
+  color: #121212;
+  font-size: 2rem;
+}
+
+.progressCercle {
+  width: 5rem;
+  height: 5rem;
+  border: 5px solid #0290ff;
+  border-radius: 50%;
+}
+
+.currentChapter {
+  display: flex;
+  flex-direction: column;
+  gap: 2rem;
+  padding-left: 5rem;
+}
+
+.progressButton {
+  height: 3.5rem;
+  width: 12rem;
+  background-color: #0290ff;
+  border: 2px solid transparent;
+  color: #fff;
+  font-size: 1rem;
+  border-radius: 12px;
+}
+
+.progressButton:hover {
+  color: #0290ff;
+  border: 2px solid #0290ff;
+  background-color: transparent;
+  cursor: pointer;
+  transition: 0.3s ease;
+}
+
+.home {
+  background-color: #fafafa;
+  padding: 3rem 5%;
+  display: flex;
+  flex-direction: column;
+  gap: 1rem;
+}
+
+@media (max-width: 768px) {
+  .bookcardcontainer {
+    grid-template-columns: 1fr;
+    grid-template-rows: 1fr 1fr 1fr;
+  }
+
+  .buttonscontainer {
+    flex-direction: column;
+  }
+
+  .bookinfocontainer {
+    display: grid;
+    grid-template-columns: 1fr 1fr;
+    gap: 2rem;
+  }
+}
+
+.navbar {
+  background-color: #fff;
+  height: 6rem;
+  padding: 0 6.25rem;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+}
+
+.links {
+  display: flex;
+  gap: 3rem;
+}
+
+.title {
+  color: #0290ff;
+  font-size: 1.875rem;
+  font-weight: bold;
+}
+
+.navList {
+  display: flex;
+  align-items: center;
+  gap: 3rem;
+}
+
+.navLink:hover {
+  opacity: 0.5;
+}
+
+.icon {
+  color: #0290ff;
+  width: 2rem;
+  height: 2rem;
+}
+
+@media (max-width: 768px) {
+  .navbar {
+    padding: 0 2rem;
+  }
+
+  .title {
+    font-size: 1.2rem;
+    text-align: center;
+  }
+
+  .icon {
+    display: none;
+  }
+}
